requirements.txt
# M&A Oracle Data Pipeline — Dependencies
# Install: pip install -r requirements.txt

# Core
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# Document parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Chunking
langchain-text-splitters>=0.0.1

# Vector store
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0

# Default embedding provider: AWS Bedrock
boto3>=1.34.0
langchain-aws>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.10

# Optional: Free local embeddings (no API key needed)
# pip install sentence-transformers langchain-huggingface

# NL2SQL service
fastapi>=0.100.0
uvicorn>=0.23.0
vanna[openai]>=0.7.0

# OPENAI langchain
#pip install langchain-openai
